pyproject.toml
[project]
name = "model-training"
version = "0.0.1"
description = "Build a model library"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "xgboost>=3.1.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
